Added different visualisation
Using the idea of FFT, and averaging in the complex plane on
exponentials (like FFT does), here's another visualisation.
Different rings are different frequencies. We start from a
frequency band setup, and use the single band coefficients to
create a new, clean data stream that we then wrap around concentric
rings. If the ring circumference matches a frequency, a
symmetric pattern will appear. The script visual_test.py
displays a randomly changing stream of the five EEG frequencies
in a slake-moth plot.